+++ CHANGELOG.md
@@ -40,7 +40,7 @@ This release named by @molxyz and [@ctrlbreak](https://twitter.com/ctrlbreak).
 - JSON API: use `\n\n` to terminate responses, for simplified parsing (pylightning now relies on this)
 - JSON API: `fundchannel` now includes an `announce` option, when false it will keep channel private. Defaults to true.
 - JSON API: `listpeers`'s `channels` now includes a `private` flag to indicate if channel is announced or not.
-- JSON API: `invoice` route hints may now include private channels if you have no public ones.
+- JSON API: `invoice` route hints may now include private channels if you have no public ones, unless new option `exposeprivatechannels` is false.
 - Plugins: experimental plugin support for `lightningd`, including option passthrough and JSON-RPC passthrough.
 
 ### Changed